Update implementation to use Context API
