+++ nuxt/index.js
@@ -1,6 +1,6 @@
 import path from 'path';
 
-export default function ExampleModule(moduleOptions) {
+export default function VsfLexascmsModule(moduleOptions) {
   this.addPlugin({
     src: path.resolve(__dirname, './plugin.js'),
     options: moduleOptions
