+++ nuxt/plugin.js
@@ -1,8 +1,8 @@
-import { setup } from 'vsf-lexascms';
+import { integrationPlugin } from 'vsf-lexascms';
 
-export default function init() {
+export default integrationPlugin(({ integration }) => {
   // Get module options
   const moduleOptions = JSON.parse('<%= JSON.stringify(options) %>');
   // Setup LexasCMS module
-  setup(moduleOptions);
-}
+  integration.configure({ ...moduleOptions });
+});
