+++ package.json
@@ -31,7 +31,7 @@
     "@types/base-64": "^0.1.3",
     "@types/jest": "^26.0.15",
     "@types/qs": "^6.9.4",
-    "@vue-storefront/core": "^2.0.12",
+    "@vue-storefront/core": "^2.2.1",
     "jest": "^26.6.3",
     "release-it": "^14.2.1",
     "release-it-lerna-changelog": "^3.1.0",
@@ -42,7 +42,7 @@
     "vue": "^2.6.12"
   },
   "peerDependencies": {
-    "@vue-storefront/core": "^2.0.10"
+    "@vue-storefront/core": "^2.2.1"
   },
   "files": [
     "dist/**/*",
