+++ src/index.test.ts
@@ -1,17 +1,19 @@
-import {
-  getConfig,
-  setup
-} from './index';
+import { createApiClient } from './index';
 
 describe('vsf-lexascms', () => {
 
-  test('setup should initialise config', async () => {
-    // Assert that config is at default value
-    expect(getConfig()).toEqual({ spaceId: null });
-    // Setup module
-    setup({ spaceId: 'space-id' });
-    // Assert that config is as expected
-    expect(getConfig()).toEqual({ spaceId: 'space-id' });
+  test('should create an api client', async () => {
+    // Create API Client
+    const apiClient = createApiClient({
+      spaceId: 'space-id'
+    });
+    // Assert
+    expect(apiClient).toMatchObject({
+      client: undefined,
+      settings: {
+        spaceId: 'space-id'
+      }
+    });
   });
 
 });
