+++ src/index.ts
@@ -1,20 +1,30 @@
+import { apiClientFactory, integrationPluginFactory } from '@vue-storefront/core';
+
+import { getContent } from './api';
 import { useContent } from './composables/useContent';
 import { LexascmsSetupConfig } from './types/lexascms';
 
-let _config: LexascmsSetupConfig = {
-  spaceId: null
-};
-
-const getConfig = (): LexascmsSetupConfig => {
-  return _config;
-};
-
+// Define setup method
 const setup = (config: LexascmsSetupConfig) => {
-  _config = config;
-};
+  return {
+    config: { ...config }
+  }
+}
+
+// Create API Client
+const { createApiClient } = apiClientFactory({
+  tag: 'lexascms',
+  onSetup: setup as any,
+  api: {
+    getContent,
+  },
+});
+
+// Create integration plugin
+const integrationPlugin = integrationPluginFactory(createApiClient);
 
 export {
-  getConfig,
-  setup,
+  createApiClient,
+  integrationPlugin,
   useContent
 };
