+++ src/types/lexascms/apiContext.ts
@@ -1,0 +1,5 @@
+import { LexascmsSetupConfig } from './setupConfig';
+
+export type LexascmsApiContext = {
+  config: LexascmsSetupConfig;
+};
