+++ src/types/lexascms/index.ts
@@ -1,3 +1,4 @@
+export * from './apiContext';
 export * from './content';
 export * from './contentSearchParams';
 export * from './requestContext';
